+++ docs/conf.py
@@ -49,11 +49,11 @@ proj_master_doc = config.get("master_doc")
 # @@@ editable
 copyright = "2023–2024, Dave Faulkmore"
 # The short X.Y.Z version.
-version = "1.4.1"
+version = "1.4.2"
 # The full version, including alpha/beta/rc tags.
-release = "1.4.1"
+release = "1.4.2"
 # The date of release, in "monthname day, year" format.
-release_date = "December 29, 2024"
+release_date = "December 30, 2024"
 # @@@ end
 
 # release = config.version